Asset Simulators

The abstract supertype AssetSimulator encompasses all simulators that can be used to generate the asset prices and underlying factors.

Each simulator has its own asset returns dynamics, these are described in the docstring of the relevant simulator (see Simulators).

Usage

In general, an AssetSimulator is passed as an input to construct a RL environment of type SimulatorEnv.

However, it is also possible to interact directly with the AssetSimulator object. In particular, the simulate_economy function can be called with the simulator as input to generate a tuple of (asset prices, asset returns, factors):

ReinforcementPortfolio.simulate_economy — Function

simulate_economy(sim, T = 500; rng = nothing)

Simulates T periods of an economy according to the dynamics of the simulator sim.

Simulators

GPSimulator

ReinforcementPortfolio.GPSimulator — Type

GPSimulator(S₀, μ_f₀, σ_f₀, B, Σ, Ψ, Φ)

An asset simulator where the prices follow a model similar to that of Gârleanu & Pedersen (2013)

The asset log returns adopt the following multidimensional factor model with $N$ assets and $K$ factors:

\[\begin{aligned} \mathbf{r}_{t+1} &= \mathbf{B}\mathbf{f}_t + \mathbf{u}_{t+1} \\ \mathbf{f}_{t+1} &= \left(\mathbf{I} - \boldsymbol{\Phi}\right) \mathbf{f}_t + \boldsymbol{\epsilon}_{t+1}. \end{aligned}\]

with

$\mathbf{B} \in \mathbb{R}^{N \times K}$ the matrix of factor loadings.

$\boldsymbol{\Phi} \in \mathbb{R}^{K \times K}$ the matrix of mean-reversion coefficients

$\mathbf{u}_t \sim \mathcal{N}(\mathbf{0}, \boldsymbol{\Sigma})$ the returns innovation term

$\boldsymbol{\epsilon}_t \sim \mathcal{N}(\mathbf{0}, \boldsymbol{\Psi})$ the factors innovation term

When an economy is simulated according to the GPSimulator, the initial factors $\mathbf{f}_0$ are randomized according to a normal distribution with mean μ_f₀ and standard deviation σ_f₀. It is assumed that the initial factor distribution is made of independent normal distributions, i.e. σ_f₀ is a vector of standard deviations and not the covariance matrix (it is the diagonal of the covariance matrix).