Simulator Environment

The SimulatorEnv integrates AssetSimulator into the generic abstract environment type AbstractEnv from ReinforcementLearning.jl.

ReinforcementPortfolio.SimulatorEnv — Type

SimulatorEnv(sim, T, fee, reward_style) <: AbstractEnv

A simulator environment, subtype of the AbstractEnv from the ReinforcementLearning.jl package.

Arguments

sim: An AssetSimulator which governs the asset returns dynamics of the SimulatorEnv.

T: The total timesteps until an episode finishes.

fee: The broker's percentage fee on each trade.

reward_style: The type of reward the agent receives when interacting with the environment.

See RewardStyle