Public Documentation

Index

ReinforcementPortfolio.BuyAndHold

ReinforcementPortfolio.ConstantlyRebalanced

ReinforcementPortfolio.PAMR

ReinforcementPortfolio.PAMR1

ReinforcementPortfolio.PAMR2

ReinforcementPortfolio.PAMRPolicy

ReinforcementPortfolio.UniformBuyAndHold

ReinforcementPortfolio.UniformCRP

Public Interface

Online Portfolio Selection Strategies

ReinforcementPortfolio.BuyAndHold — Type

BuyAndHold <: AbstractPolicy

Buy and hold strategy. Buy some amount of stock w and hold them forever.

ReinforcementPortfolio.UniformBuyAndHold — Function

UniformBuyAndHold(env)

Creates a uniform BuyAndHold strategy for the given environment env.

ReinforcementPortfolio.ConstantlyRebalanced — Type

ConstantlyRebalanced <: AbstractPolicy

Constantly rebalanced portfolio. Reallocates to fixed weights w at chosen frequency freq.

ReinforcementPortfolio.UniformCRP — Function

UniformCRP(env, freq = 30)

Creates a uniform ConstantlyRebalanced strategy for the given environment env and rebalancing frequency freq.

ReinforcementPortfolio.PAMRPolicy — Type

PAMRPolicy

Abstract type for Passive Aggressive Mean Reversion strategies.

Reference: B. Li, P. Zhao, S. C.H. Hoi, and V. Gopalkrishnan. Pamr: Passive aggressive mean reversion strategy for portfolio selection, 2012. https://link.springer.com/content/pdf/10.1007/s10994-012-5281-z.pdf

ReinforcementPortfolio.PAMR — Type

PAMR(w, ϵ, freq = 30)

Original Passive Aggressive Mean Reversion strategy. See PAMRPolicy.

ReinforcementPortfolio.PAMR1 — Type

PAMR1(w, ϵ, C, freq = 30)

First variant of the Passive Aggressive Mean Reversion strategy. See PAMRPolicy.

ReinforcementPortfolio.PAMR2 — Type

PAMR2(w, ϵ, C, freq = 30)

Second variant of the Passive Aggressive Mean Reversion strategy. See PAMRPolicy.